+++ src/Cortside.Common.Hosting.Tests/TimedHostedServiceTests.cs
@@ -45,7 +45,7 @@ namespace Cortside.Common.Hosting.Tests {
 #pragma warning disable CS4014 // Because this call is not awaited, execution of the current method continues before the call is completed
             instance.PublicExecuteAsync(stoppingToken);
 #pragma warning restore CS4014 // Because this call is not awaited, execution of the current method continues before the call is completed
-            await Task.Delay(1000).ConfigureAwait(false);
+            await Task.Delay(1000);
 
             // Assert
             Assert.True(instance.Executed);