+++ CHANGELOG.md
@@ -1,3 +1,54 @@
+# Release 6.1
+
+* Update nuget dependencies to latest stable versions
+* Add ScopedLocalTimeZone for use with tests to set predictable timezone regardless of environment
+    ```csharp
+    using (new ScopedLocalTimeZone(TimeZoneInfo.FindSystemTimeZoneById("UTC+12"))) {
+        var localDateTime = new DateTime(2020, 12, 31, 23, 59, 59, DateTimeKind.Local);
+        var utcDateTime = TimeZoneInfo.ConvertTimeToUtc(localDateTime);
+
+        Assert.Equal("UTC+12", TimeZoneInfo.Local.Id);
+        Assert.Equal(utcDateTime.AddHours(12), localDateTime);
+    }
+    ```
+* Add XunitLogger for capturing log output to xUnit's ITestOutputHelper
+	```csharp
+	// Create a logger factory with a debug provider
+	loggerFactory = LoggerFactory.Create(builder => {
+		builder
+			.SetMinimumLevel(LogLevel.Trace)
+			.AddFilter("Microsoft", LogLevel.Warning)
+			.AddFilter("System", LogLevel.Warning)
+			.AddFilter("Cortside.Common", LogLevel.Trace)
+			.AddXunit(output);
+	});
+
+	// Create a logger with the category name of the current class
+	var logger = loggerFactory.CreateLogger<XunitLoggerTest>();
+
+	... user logger as you would any other logger ...
+	```
+* changed hierarchichal organization of loggers, so LogEventLogger has new namespace of Cortside.Common.Testing.Logging.LogEvent
+* Added helper class RandomValues for generating "random" data
+* Added IServiceCollection extension method Unregister<T> for unregistering something already in service collection
+	```csharp
+	// Remove the app's DbContext registration.
+	services.Unregister<DbContextOptions<DatabaseContext>>();
+	services.Unregister<DbContext>();
+	```
+
+|Commit|Date|Author|Message|
+|---|---|---|---|
+| 9b1a714 | <span style="white-space:nowrap;">2023-08-29</span> | <span style="white-space:nowrap;">Cort Schaefer</span> |  update version
+| 1007bdd | <span style="white-space:nowrap;">2023-08-29</span> | <span style="white-space:nowrap;">Cort Schaefer</span> |  update release notes
+| f0f453f | <span style="white-space:nowrap;">2023-09-04</span> | <span style="white-space:nowrap;">Cort Schaefer</span> |  Merge branch 'master' into develop
+| 3c480ba | <span style="white-space:nowrap;">2023-09-12</span> | <span style="white-space:nowrap;">Cort Schaefer</span> |  add new project for common unit and integration test related classes
+| a11f4cd | <span style="white-space:nowrap;">2023-10-16</span> | <span style="white-space:nowrap;">Cort Schaefer</span> |  Add LogEventLogger to help in validating and testing log output
+| e2f06ca | <span style="white-space:nowrap;">2023-10-31</span> | <span style="white-space:nowrap;">Cort Schaefer</span> |  add extension method for unregistering a service from service collection
+| 184c4a2 | <span style="white-space:nowrap;">2023-11-08</span> | <span style="white-space:nowrap;">Cort Schaefer</span> |  add new XunitLogger for capturing ITestOutputHelper output
+| 5efb5a7 | <span style="white-space:nowrap;">2023-11-08</span> | <span style="white-space:nowrap;">Cort Schaefer</span> |  (HEAD -> develop, origin/develop, origin/HEAD) update to latest nuget packages
+****
+
 # Release 6.0
 
 * Update version number to match framework version (6.x)